requirements.txt
python-telegram-bot==21.4
google-generativeai==0.8.3
pandas==2.2.3
openpyxl==3.1.5